CLN/TST Remove compat.scipy
